requirements.txt
Django=4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-model-utils==4.3.1
mysqlclient==2.2.0